Remove redondant cast
Former-commit-id: 4daac8582d32b2eb0bcca31400d2cf8c4cd6e0df [formerly ccdfc8d5ac7efa5806a68c9b6d082ec86212bdf6] [formerly 9eca950d39cc4301be1bd0e0ff6e6a0ef8b6cc94 [formerly 2ec2386967389c5b74c3d838d956b7e6e80f45e8] [formerly 67d8f7c2015e47816b7db43ff5fbd1e9cc9579cb] [formerly 5a6e9a44de0fa1c5525907586257e5f8233710a1 [formerly b5e8d02cfc5165f938b37ede021606e5bdf444aa]] [formerly 9eca950d39cc4301be1bd0e0ff6e6a0ef8b6cc94 [formerly 2ec2386967389c5b74c3d838d956b7e6e80f45e8] [formerly 67d8f7c2015e47816b7db43ff5fbd1e9cc9579cb] [formerly 5a6e9a44de0fa1c5525907586257e5f8233710a1 [formerly b5e8d02cfc5165f938b37ede021606e5bdf444aa]] [formerly d3715faeb524ef77a75cd850e18be8e6341d8b25 [formerly 09d11c3ebbd27d61c66b5486aca41efd107ba8b8]]]]
Former-commit-id: 9c2af78c3c6e23d8c39cd4c885417e3930a1dc2a
Former-commit-id: 271b37433378f27320ef2bd66405170fe16687d0
Former-commit-id: 2f5db4666ac572496248884472e73c7a8dd7b32a
Former-commit-id: 6b073b3e7fcc3518859c8ca7c11c620a8ed849fc
Former-commit-id: af7c448cb9adeaefe05a33e39d9153f509ed1596 [formerly 28ea981da1ab74401dfc497366f17b1ab359f5a0]
Former-commit-id: 1838c76a0ac85867a2cfca1759def319e644bdde
Former-commit-id: 0c4325932442b4182ce7f3046ec7780625e34216

+++ Source/Raw.Net/Source/Decoder/DNGDecoder.cs
@@ -129,8 +129,8 @@ namespace RawNet
 
             for (UInt32 i = 0; i < nrects; i++)
             {
-                iPoint2D topleft = new iPoint2D((int)rects[i * 4 + 1], (int)rects[i * 4]);
-                iPoint2D bottomright = new iPoint2D((int)rects[i * 4 + 3], (int)rects[i * 4 + 2]);
+                iPoint2D topleft = new iPoint2D(rects[i * 4 + 1], rects[i * 4]);
+                iPoint2D bottomright = new iPoint2D(rects[i * 4 + 3], rects[i * 4 + 2]);
                 // Is this a horizontal box, only add it if it covers the active width of the image
                 if (topleft.x <= top.x && bottomright.x >= (mRaw.dim.x + top.x))
                     mRaw.blackAreas.Add(new BlackArea(topleft.y, bottomright.y - topleft.y, false));
@@ -199,7 +199,7 @@ namespace RawNet
                     black_sum[i & 1] += blackleveldeltav.getFloat(i);
 
                 for (int i = 0; i < 4; i++)
-                    mRaw.blackLevelSeparate[i] += (int)(black_sum[i >> 1] / (float)mRaw.dim.y * 2.0f);
+                    mRaw.blackLevelSeparate[i] += (int)(black_sum[i >> 1] / mRaw.dim.y * 2.0f);
             }
 
             if (raw.hasEntry(TagType.BLACKLEVELDELTAH))
@@ -212,7 +212,7 @@ namespace RawNet
                     black_sum[i & 1] += blackleveldeltah.getFloat(i);
 
                 for (int i = 0; i < 4; i++)
-                    mRaw.blackLevelSeparate[i] += (int)(black_sum[i & 1] / (float)mRaw.dim.x * 2.0f);
+                    mRaw.blackLevelSeparate[i] += (int)(black_sum[i & 1] / mRaw.dim.x * 2.0f);
             }
             return true;
         }
